rename CSS modules to SCSS for Button, PollComponent, and VotersList
