+++ frontend/src/Components/Poll/PollComponent.module.scss
