+++ frontend/src/Components/Poll/PollComponent.tsx
@@ -10,7 +10,7 @@ import Radio from '@ui/Radio'
 import { pluralize } from '@utils/utils'
 import { toast } from 'react-toastify'
 
-import styles from './PollComponent.module.css'
+import styles from './PollComponent.module.scss'
 
 interface PollProps {
   pollId: number
