+++ frontend/src/Components/Poll/VotersList.module.scss
