+++ frontend/src/Components/UI/Button.tsx
@@ -2,7 +2,7 @@ import React, { ButtonHTMLAttributes } from 'react'
 
 import classNames from 'classnames'
 
-import styles from './Button.module.css'
+import styles from './Button.module.scss'
 
 /**
  * Interface for Button component props
